🚛 copy cluster into components/
